Reimplemented sessions
Added sessions global variable to hold all user sessions.
A user session has the sessionId as the key and tokens/user info
as the value.
There is a guest session object to be used for queries by
users that don't login to Twitter.
There is GET /checkSession to see if a sessionId is valid and
not the guest sessionId.
There is GET /guestSession that returns the guest sessionId.